refactor: extract asyncio subprocess filter from tui to core for MCP reuse
