+++ src/kagan/tui/textual_compat.py
@@ -1,13 +1,14 @@
 """Compatibility workarounds for Textual versions used by Kagan."""
 
-import asyncio
 from typing import Any
 
 from loguru import logger
 from textual.widgets._select import SelectOverlay
 
+# Re-export from core so existing TUI imports continue to work.
+from kagan.core._asyncio_compat import install_asyncio_subprocess_exception_filter
+
 _PATCH_FLAG = "_kagan_select_overlay_move_page_patched"
-_ASYNCIO_SUBPROCESS_HANDLER_PATCH_FLAG = "_kagan_asyncio_subprocess_handler_patched"
 
 
 def _patch_select_overlay_page_navigation() -> None:
@@ -48,46 +49,6 @@ def _patch_select_overlay_page_navigation() -> None:
 
     setattr(_safe_move_page, _PATCH_FLAG, True)
     SelectOverlay._move_page = _safe_move_page
-
-
-def _is_known_asyncio_subprocess_invalid_state(context: dict[str, Any]) -> bool:
-    exc = context.get("exception")
-    message = str(context.get("message") or "")
-    if isinstance(exc, RuntimeError) and "Event loop is closed" in str(exc):
-        if "SubprocessTransport" in message:
-            return True
-    if not isinstance(exc, asyncio.InvalidStateError):
-        return False
-    handle = context.get("handle")
-    callback = getattr(handle, "_callback", None)
-    callback_name = str(getattr(callback, "__qualname__", ""))
-    if "_call_connection_lost" not in f"{callback_name} {message}":
-        return False
-    return "BaseSubprocessTransport" in message or "_UnixReadPipeTransport" in message
-
-
-def install_asyncio_subprocess_exception_filter(
-    loop: asyncio.AbstractEventLoop | None = None,
-) -> None:
-    target = loop or asyncio.get_running_loop()
-    if getattr(target, _ASYNCIO_SUBPROCESS_HANDLER_PATCH_FLAG, False):
-        return
-    previous_handler = target.get_exception_handler()
-
-    def _handler(active_loop: asyncio.AbstractEventLoop, context: dict[str, Any]) -> None:
-        if _is_known_asyncio_subprocess_invalid_state(context):
-            logger.debug(
-                "Ignoring known asyncio subprocess shutdown race: {}",
-                context.get("message") or "InvalidStateError in _call_connection_lost",
-            )
-            return
-        if previous_handler is not None:
-            previous_handler(active_loop, context)
-            return
-        active_loop.default_exception_handler(context)
-
-    target.set_exception_handler(_handler)
-    setattr(target, _ASYNCIO_SUBPROCESS_HANDLER_PATCH_FLAG, True)
 
 
 def apply_textual_compat_workarounds() -> None:
